Update: - add resource creator for fast creating items and add his to DB - remove cost from items add value - more)

+++ src/game/classes/items/damages/lightning_dmg.py
@@ -1,6 +1,6 @@
 class LightningDamage:
-    def __init__(self, value: int) -> None:
+    def __init__(self, value: float) -> None:
         self.value = value
 
-    def get_value(self) -> int:
+    def get_value(self) -> float:
         return self.value
