+++ src/game/classes/items/damages/physical_dmg.py
@@ -1,6 +1,6 @@
 class PhysicalDamage:
-    def __init__(self, value: int) -> None:
+    def __init__(self, value: float) -> None:
         self.value = value
 
-    def get_value(self) -> int:
+    def get_value(self) -> float:
         return self.value
